requirements.txt
customtkinter>=5.2.0
faster-whisper>=1.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
pydub>=0.25.1
numpy>=1.24.0
pillow>=10.0.0
# LLM providers for feedback organization
ollama>=0.1.0
openai>=1.0.0
anthropic>=0.18.0
# Export functionality
reportlab>=4.0.0
python-docx>=1.0.0
